Renamed utils to super_matrix and started adding docs
