+++ src/c/comp_col.rs
@@ -1,4 +1,4 @@
-use crate::c::utils::{c_SuperMatrix, Dtype_t, Mtype_t, Stype_t};
+use crate::c::super_matrix::{c_SuperMatrix, Dtype_t, Mtype_t, Stype_t};
 use libc;
 
 #[link(name = "superlu")]
