+++ src/c/expert_driver.rs
@@ -1,6 +1,6 @@
 use crate::c::options::superlu_options_t;
 use crate::c::stat::SuperLUStat_t;
-use crate::c::utils::c_SuperMatrix;
+use crate::c::super_matrix::c_SuperMatrix;
 use libc;
 
 #[repr(C)]
