+++ src/super_matrix.rs
@@ -1,4 +1,4 @@
-use crate::c::utils::c_SuperMatrix;
+use crate::c::super_matrix::c_SuperMatrix;
 
 pub trait SuperMatrix {
     fn super_matrix<'a>(&'a mut self) -> &mut c_SuperMatrix;
